Fix dialyzer spec issues
Removed the underspecs dialyzer option which was causing dialyzer to
exit early due to a known dialyzer bug. This revealed an invalid type
spec for sign_request. The return type shouldn't need to be tuples of
`{[any()], [any()]}`, but tracking that down at this time may do more
harm than good.