Delay start until keygen_start_size keys are available
New app config `{chef_authn, keygen_start_size, Int}` can be used to
control how many keys should be available when the cache is
initialized. The cache's startup will block until the specified number
of keys are created.

Also included is a bit of reorganization of the tests with a new make
target `make slow_tests`. To get test coverage for timeout and worker
crash cases and observe replenishment of cache takes a bit of time and
relies on speed of system. These tests may need tuning to pass on
other hardware.