Parallelize key creation in chef_keygen_cache
This patch introduces a new app config:

     `{chef_authn, keygen_cache_workers, WorkerCount}`

that controls the number of concurrent keygen workers that will be
used to fill the cache to its max level. The default worker count is
set to half of the number of logical processors available on the
system.

The value returned when a worker times out changed from 'timeout' to
'keygen_timeout' to distinguish it from gen_server system timeouts.

keygen constants moved to chef_keygen.hrl