Add multiple key options to authn sequence
To support key rotation we would like to try to auth against a list of
keys and return the first key that succeeds. This extends the
authenticate_user_request API to accept a list of {key_description,
public_key} pairs and returns {name, user, and key_description} if one
matches.