Refactor verify sig code
I'd like to be able to allow different types of algorithms to be
used for signing v1.3 such as sha256, sha512, etc. This change
passes that information down to the appropriate functions.

While I had to change the tests, no public interfaces have been
changed. They required changing because during testing, all
functions are exported so they can be tested