Do not hash path and userid in the 1.3 signature format
Hashing these things is no longer required. The reason
they were hashing in v1.0/v1.1 is because the message
was directly signed. This meant that there was a limit on
the size of the message.

v1.3 does not have this requirement as RSASSA-PKCS#1-v1_5
signs a hash of the message.