Handle already-decoded key data in all signing versions
The decrypt_sig key handled both decoded (i.e. values already of type
rsa_public_key()) and binary data.

Unfortunately, it appears that only the verification code for signing
versions 1.0 and 1.1 call decrypt_sig.

This PR changes the commonly used decode_key_data to handle the case
of already decoded data.

I haven't fully bisected this to understand how long this has been
broken for, but my guess is that the user-facing breakage was
potentially introduced because of a change in the underlying
public_key functions on different erlang versions.

Signed-off-by: Steven Danna <steve@chef.io>